debugging external list
