+++ src/components/app/list/external.util.tsx
@@ -1,6 +1,7 @@
-import { components } from 'react-select';
-import { ReactComponent as ArrowDown } from '../../../assets/icons/ic-chevron-down.svg';
 import React from 'react';
+import Select, { components } from 'react-select';
+import { ReactComponent as ArrowDown } from '../../../assets/icons/ic-chevron-down.svg';
+import { Option, multiSelectStyles } from '../../common';
 
 export const ValueContainer = props => {
     let length = props.getValue().length;
@@ -33,4 +34,61 @@ export const DropdownIndicator = props => {
             <ArrowDown className={`rotate fcn-4`} style={{ ['--rotateBy' as any]: props.selectProps.menuIsOpen ? '180deg' : '0deg', height: '20px', width: '20px' }} />
         </components.DropdownIndicator>
     )
+}
+
+export function ExternalFilters({ handleSelectedCluster, handleSelectedNamespace, cluster, namespace, selectedNamespace, setNamespace, selectedCluster, setCluster }) {
+    const MenuList = (props) => {
+        let name = props.selectProps.name
+        return (
+            <components.MenuList {...props}>
+                {props.children}
+                <div className="chartListApplyFilter flex bcn-0 pt-10 pb-10">
+                    <button type="button" style={{ width: "92%" }} className="cta flex cta--chart-store"
+                        disabled={false}
+                        onClick={(event) => { name === "cluster" ? handleSelectedCluster(event) : handleSelectedNamespace(event) }}
+                    >Apply Filter</button>
+                </div>
+            </components.MenuList>
+        );
+    };
+    return <>
+        <Select className="cn-9 fs-13"
+            placeholder="Cluster: All"
+            name="cluster"
+            options={cluster}
+            components={{
+                Option,
+                MenuList,
+                ValueContainer,
+                DropdownIndicator,
+                IndicatorSeparator: null,
+            }}
+            onChange={(selected: any) => setCluster(selected)}
+            isMulti
+            value={selectedCluster}
+            hideSelectedOptions={false}
+            closeMenuOnSelect={false}
+            isClearable={false}
+            styles={multiSelectStyles}
+        />
+        <Select className="cn-9 fs-14"
+            placeholder="Namespace: All"
+            options={namespace}
+            onChange={(selected: any) => setNamespace(selected)}
+            value={selectedNamespace}
+            name="namespace"
+            components={{
+                Option,
+                MenuList,
+                ValueContainer,
+                IndicatorSeparator: null,
+                DropdownIndicator,
+            }}
+            isClearable={false}
+            isMulti
+            hideSelectedOptions={false}
+            closeMenuOnSelect={false}
+            styles={{ ...multiSelectStyles }}
+        />
+    </>
 }
